Build with docker and sh rather than nix
After this change, all linux packages will be built inside docker
containers and macos packages will be built with shell scripts that run
directly in the github action. The main reason is that it allows us to
use opam for package management rather than nix, which means that newly
released packages become available the instant they are released, rather
than needing to wait for new versions to land in nix's package repo.

Another issue that this solves is that some macos builds of
ocaml-lsp-server don't work properly when built with nix (read more
[here](https://discuss.ocaml.org/t/corrupted-compiled-interface-after-dune-build/14919)).

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>